refactor: extract math helpers
